tweet_bot/requirements.txt
requests
tweepy
yfinance
beautifulsoup4
Pillow
python-dotenv
